% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_course}
\alias{install_course}
\title{Install workshop materials for the Applied Biostatistics Course with R.}
\usage{
install_course(destdir = getOption("usethis.destdir"))
}
\arguments{
\item{destdir}{The path on your computer where you would like the workshop
installed. By default, this will install somewhere conspicuous, like your
desktop, although you can tell `install_course()` exactly where you want
it to download.}
}
\description{
`install_course()` will install the workshop materials for the workshop in a new repository.
}
